+++ src/providers/fractal/hooks/useFractal.ts
@@ -1,9 +1,10 @@
 import { createContext, useContext, Context } from 'react';
-import { FractalDAO, IDaoLegacy } from '../types';
+import { FractalActions, FractalDAO, IDaoLegacy } from '../types';
 
 export interface IFractalContext {
   dao: FractalDAO;
   daoLegacy: IDaoLegacy;
+  dispatch: React.Dispatch<FractalActions>;
 }
 
 export const FractalContext = createContext<IFractalContext | null>(null);
